feat: Allow registering the listener synchronously
Instead of scheduling listener registration for onMounted()
one can now also directly register the listener (using a new
'immediate' option). This helps with setting up the listener
after mounting, because in that case onMounted() does
not run.

Also: Using onScopeDispose() over onBeforeUnmount().
This does not behave differently for the usual setup case
but it might help in explicit scope cases.